Move shareable code and types out of PurchasesTable
