+++ packages/web/src/pages/purchases-page/PurchasesTable.module.css
@@ -14,22 +14,3 @@
   fill: var(--neutral-light-4);
 }
 
-.contentName {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--unit-6);
-}
-
-.artwork {
-  height: var(--unit-10);
-  width: var(--unit-10);
-  border-radius: var(--unit-1);
-  border: 1px solid var(--border-default);
-  overflow: hidden;
-}
-
-.artistName {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--unit-1);
-}
