+++ packages/web/src/pages/purchases-page/components/TrackNameWithArtwork.module.css
@@ -1,0 +1,13 @@
+.container {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--unit-6);
+}
+
+.artwork {
+  height: var(--unit-10);
+  width: var(--unit-10);
+  border-radius: var(--unit-1);
+  border: 1px solid var(--border-default);
+  overflow: hidden;
+}
