+++ packages/web/src/pages/purchases-page/components/UserNameWithBadges.module.css
@@ -1,0 +1,5 @@
+.container {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--unit-1);
+}
