+++ packages/web/src/pages/purchases-page/types.ts
@@ -1,0 +1,5 @@
+import { USDCPurchaseDetails } from '@audius/common'
+import { Cell, Row } from 'react-table'
+
+export type PurchaseCell = Cell<USDCPurchaseDetails>
+export type PurchaseRow = Row<USDCPurchaseDetails>
