+++ packages/web/src/pages/purchases-page/utils.ts
@@ -1,0 +1,5 @@
+import { PurchaseRow } from './types'
+
+export const isEmptyRow = (row?: PurchaseRow) => {
+  return !row?.original.signature
+}
